+++ examples/service/api/call_feedback_summary/create/feedback_summary_create_example.go
@@ -1,0 +1,42 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v2010 "github.com/RJPearson94/twilio-sdk-go/service/api/v2010"
+	"github.com/RJPearson94/twilio-sdk-go/service/api/v2010/account/calls/feedback_summaries"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var apiSession *v2010.V2010
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	apiSession = twilio.NewWithCredentials(creds).API.V2010
+}
+
+func main() {
+	resp, err := apiSession.
+		Account("ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		Calls.
+		FeedbackSummaries.
+		Create(&feedback_summaries.CreateFeedbackSummaryInput{
+			StartDate: "2019-10-03",
+			EndDate:   "2020-10-03",
+		})
+
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	log.Printf("SID: %s", resp.Sid)
+}