+++ service/api/v2010/account/calls/api_op_client.go
@@ -1,7 +1,11 @@
 // Package calls contains auto-generated files. DO NOT MODIFY
 package calls
 
-import "github.com/RJPearson94/twilio-sdk-go/client"
+import (
+	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/api/v2010/account/calls/feedback_summaries"
+	"github.com/RJPearson94/twilio-sdk-go/service/api/v2010/account/calls/feedback_summary"
+)
 
 // Client for managing call resources
 // See https://www.twilio.com/docs/voice/api/call-resource for more details
@@ -9,6 +13,9 @@ type Client struct {
 	client *client.Client
 
 	accountSid string
+
+	FeedbackSummaries *feedback_summaries.Client
+	FeedbackSummary   func(string) *feedback_summary.Client
 }
 
 // ClientProperties are the properties required to manage the calls resources
@@ -22,5 +29,15 @@ func New(client *client.Client, properties ClientProperties) *Client {
 		client: client,
 
 		accountSid: properties.AccountSid,
+
+		FeedbackSummaries: feedback_summaries.New(client, feedback_summaries.ClientProperties{
+			AccountSid: properties.AccountSid,
+		}),
+		FeedbackSummary: func(feedbackSummarySid string) *feedback_summary.Client {
+			return feedback_summary.New(client, feedback_summary.ClientProperties{
+				AccountSid: properties.AccountSid,
+				Sid:        feedbackSummarySid,
+			})
+		},
 	}
 }